+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.9
 - Anvil grid scales with scene length unit
+- Fix initial texture application setting incorrect UV scale
 
 ## 1.2.8
 - Swapped Level Design and Hotspot Mapping workspaces