+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix initial texture application setting incorrect UV scale
 - Improve reliability of image texture preview
 - Debug logging is now a scene property
+- Texture preview shows previous image selected (disabled)
 
 ## 1.2.8
 - Swapped Level Design and Hotspot Mapping workspaces