+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Debug logging is now a scene property
 - Texture preview shows previous image selected (disabled)
 - New boxes from box builder are textured with the previously selected texture
+- Selecting a texture in the file browser will apply it to an entire object while in object mode
 
 ## 1.2.8
 - Swapped Level Design and Hotspot Mapping workspaces